% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chen_logistic.R
\name{chen_logistic}
\alias{chen_logistic}
\title{Chen & Chen Logistic}
\usage{
chen_logistic(X_l, Y_l, f_l, X_u, f_u, intercept = TRUE)
}
\arguments{
\item{X_l}{(matrix): n x p matrix of covariates in the labeled data.}

\item{Y_l}{(vector): n-vector of labeled outcomes.}

\item{f_l}{(vector): n-vector of predictions in the labeled data.}

\item{X_u}{(matrix): N x p matrix of covariates in the unlabeled data.}

\item{f_u}{(vector): N-vector of predictions in the unlabeled data.}

\item{intercept}{(Logical): Do the design matrices include intercept
columns? Default is \code{TRUE}.}
}
\value{
(list): A list containing the following:
\describe{
\item{est}{(vector): vector of Chen & Chen logistic regression coefficient
estimates.}
\item{se}{(vector): vector of standard errors of the coefficients.}
}
}
\description{
Helper function for Chen & Chen logistic regression estimation.
}
\details{
Another look at statistical inference with machine learning-imputed data
(Gronsbell et al., 2026) \doi{10.48550/arXiv.2411.19908}
}
\examples{

dat <- simdat(model = "logistic")

form <- Y - f ~ X1

X_l <- model.matrix(form, data = dat[dat$set_label == "labeled", ])

Y_l <- dat[dat$set_label == "labeled", all.vars(form)[1]] |>
  matrix(ncol = 1)

f_l <- dat[dat$set_label == "labeled", all.vars(form)[2]] |>
  matrix(ncol = 1)

X_u <- model.matrix(form, data = dat[dat$set_label == "unlabeled", ])

f_u <- dat[dat$set_label == "unlabeled", all.vars(form)[2]] |>
  matrix(ncol = 1)

chen_logistic(X_l, Y_l, f_l, X_u, f_u, intercept = TRUE)

}
