% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chen_ols.R
\name{chen_ols}
\alias{chen_ols}
\title{Chen & Chen OLS}
\usage{
chen_ols(Y_l, X_l, f_l, f_u, X_u)
}
\arguments{
\item{Y_l}{(vector): n-vector of labeled outcomes.}

\item{X_l}{(matrix): n x p matrix of covariates in the labeled data.}

\item{f_l}{(vector): n-vector of predictions in the labeled data.}

\item{f_u}{(vector): N-vector of predictions in the unlabeled data.}

\item{X_u}{(matrix): N x p matrix of covariates in the unlabeled data.}
}
\value{
(list): A list containing the following:

\describe{
\item{est}{(vector): vector of Chen & Chen OLS regression coefficient
estimates.}
\item{se}{(vector): vector of standard errors of the coefficients.}
}
}
\description{
Helper function for Chen & Chen OLS estimation
}
\details{
Another look at inference after prediction (Gronsbell et al., 2025)
\url{https://arxiv.org/pdf/2411.19908}
}
\examples{

dat <- simdat(model = "ols")

form <- Y - f ~ X1

X_l <- model.matrix(form, data = dat[dat$set_label == "labeled", ])

Y_l <- dat[dat$set_label == "labeled", all.vars(form)[1]] |>
  matrix(ncol = 1)

f_l <- dat[dat$set_label == "labeled", all.vars(form)[2]] |>
  matrix(ncol = 1)

X_u <- model.matrix(form, data = dat[dat$set_label == "unlabeled", ])

f_u <- dat[dat$set_label == "unlabeled", all.vars(form)[2]] |>
  matrix(ncol = 1)

#chen_ols(X_l, Y_l, f_l, X_u, f_u)

}
