% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_quantile.R
\name{compute_cdf_diff}
\alias{compute_cdf_diff}
\title{Empirical CDF Difference}
\usage{
compute_cdf_diff(Y, Yhat, grid, w = NULL)
}
\arguments{
\item{Y}{(matrix): n x 1 matrix of observed data.}

\item{Yhat}{(matrix): n x 1 matrix of predictions.}

\item{grid}{(matrix): Grid of values to compute the CDF at.}

\item{w}{(vector, optional): n-vector of sample weights.}
}
\value{
(list): Difference between the empirical CDFs of the data and the
predictions and its standard deviation at the specified grid points.
}
\description{
Computes the difference between the empirical CDFs of the data and the
predictions.
}
\examples{
# Need example code

}
