% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ipd.R
\name{.build_design}
\alias{.build_design}
\title{Build design matrices and outcome vectors}
\usage{
.build_design(formula, data_l, data_u, intercept = TRUE, na_action = "na.fail")
}
\arguments{
\item{formula}{two-sided formula \code{Y - f ~ X...}}

\item{data_l}{labeled data.frame}

\item{data_u}{unlabeled data.frame}

\item{intercept}{include intercept?}

\item{na_action}{"na.fail" or "na.omit"}
}
\value{
list(X_l, Y_l, f_l, X_u, f_u)
}
\description{
Build design matrices and outcome vectors
}
\keyword{internal}
