% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pspa.R
\name{link_grad}
\alias{link_grad}
\title{Gradient of the link function}
\usage{
link_grad(t, method = c("ols", "logistic", "poisson"))
}
\arguments{
\item{t}{t}

\item{method}{indicates the method to be used for M-estimation.
Options include "mean", "quantile", "ols", "logistic", and "poisson".}
}
\value{
gradient of the link function
}
\description{
\code{link_grad} function for gradient of the link function
}
