% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pspa.R
\name{link_Hessian}
\alias{link_Hessian}
\title{Hessians of the link function}
\usage{
link_Hessian(t, method = c("logistic", "poisson"))
}
\arguments{
\item{t}{t}

\item{method}{indicates the method to be used for M-estimation.
Options include "mean", "quantile", "ols", "logistic", and "poisson".}
}
\value{
Hessians of the link function
}
\description{
\code{link_Hessian} function for Hessians of the link function
}
