% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mean_psi}
\alias{mean_psi}
\title{Sample expectation of psi}
\usage{
mean_psi(X, Y, theta, quant = NA, method)
}
\arguments{
\item{X}{Array or data.frame containing covariates}

\item{Y}{Array or data.frame of outcomes}

\item{theta}{parameter theta}

\item{quant}{quantile for quantile estimation}

\item{method}{indicates the method to be used for M-estimation.
Options include "mean", "quantile", "ols", "logistic", and "poisson".}
}
\value{
sample expectation of psi
}
\description{
\code{mean_psi} function for sample expectation of psi
}
