% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdc_ols.R
\name{pdc_ols}
\alias{pdc_ols}
\title{PDC OLS}
\usage{
pdc_ols(X_l, Y_l, f_l, X_u, f_u, intercept = TRUE)
}
\arguments{
\item{X_l}{(matrix): n x p matrix of covariates in the labeled data.}

\item{Y_l}{(vector): n-vector of labeled outcomes.}

\item{f_l}{(vector): n-vector of predictions in the labeled data.}

\item{X_u}{(matrix): N x p matrix of covariates in the unlabeled data.}

\item{f_u}{(vector): N-vector of predictions in the unlabeled data.}

\item{intercept}{(Logical): Do the design matrices include intercept
columns? Default is \code{TRUE}.}
}
\value{
(list): A list containing the following:
\describe{
\item{est}{(vector): vector of PDC OLS regression coefficient estimates.}
\item{se}{(vector): vector of standard errors of the coefficients.}
}
}
\description{
Helper function for PDC OLS estimation.
}
\details{
Prediction de-correlated inference: A safe approach for post-prediction
inference (Gan et al., 2024) \doi{10.1111/anzs.12429}
}
\examples{

dat <- simdat(model = "ols")

form <- Y - f ~ X1

X_l <- model.matrix(form, data = dat[dat$set_label == "labeled", ])

Y_l <- dat[dat$set_label == "labeled", all.vars(form)[1]] |>
  matrix(ncol = 1)

f_l <- dat[dat$set_label == "labeled", all.vars(form)[2]] |>
  matrix(ncol = 1)

X_u <- model.matrix(form, data = dat[dat$set_label == "unlabeled", ])

f_u <- dat[dat$set_label == "unlabeled", all.vars(form)[2]] |>
  matrix(ncol = 1)

pdc_ols(X_l, Y_l, f_l, X_u, f_u, intercept = TRUE)

}
