% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popinf_ols.R
\name{popinf_ols}
\alias{popinf_ols}
\title{Assumption-lean and data-adaptive post-prediction inference (Miao et al. 2023)}
\usage{
popinf_ols(
  rec_form,
  inf_form,
  dat,
  alpha = 0.05,
  eta = 0.001,
  delta = 1e-04,
  K = 100
)
}
\arguments{
\item{rec_form}{A formula defining the rectifier model. This should be of
the form Y - Yhat ~ X, where Y is the name of the column corresponding to
the observed outcome in the labeled data, Yhat is the name of the column
corresponding to the predicted outcome in the labeled data, and X generally
corresponds to the features of interest (e.g., X1 + X2).}

\item{inf_form}{A formula defining the inference model. This should be of
the form Yhat ~ X, where Yhat is the name of the column corresponding to the
predicted outcome in the unlabeled data, and X generally corresponds to the
features of interest (e.g., X1 + X2).}

\item{dat}{data in the form of the simdat function}

\item{alpha}{scalar type I error rate for hypothesis testing - values in (0, 1); defaults to 0.05}

\item{eta}{scalar step size for gradient descent; defaults to 0.001}

\item{delta}{scalar tolerance for assessing convergence; defaults to 0.0001}

\item{K}{integer maximum number of iterations; defaults to 100}
}
\value{
A list of outputs: estimate of inference model parameters and corresponding standard error.
}
\description{
A short description...
}
\details{
Additional details...
}
\examples{

rec_form <- Y - Yhat ~ X1

inf_form <- Yhat ~ X1

dat <- simdat()

popinf_ols(rec_form, inf_form, dat = dat)

}
