% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postpi_boot_ols.R
\name{postpi_boot_ols}
\alias{postpi_boot_ols}
\title{IPD Linear Regression using Wang et al. (2020) Bootstrap Correction}
\usage{
postpi_boot_ols(rel_form, inf_form, dat, nboot = 100, rel_func = "lm")
}
\arguments{
\item{rel_form}{A formula defining the relationship model. This should be
of the form Y ~ Yhat, where Y is the name of the column corresponding to
the observed outcome in the labeled data and Yhat is the name of the column
corresponding to the predicted outcome in the labeled data.}

\item{inf_form}{A formula defining the inference model. This should be of
the form Yhat ~ X, where Yhat is the name of the column corresponding to the
predicted outcome in the unlabeled data, and X generally corresponds to the
features of interest (e.g., X1 + X2).}

\item{dat}{data in the form of the simdat function}

\item{nboot}{number of bootstraps}

\item{rel_func}{relationship function model form (lm, gam, glm, etc)}
}
\value{
A list of outputs: estimate of inference model parameters and corresponding standard error based on both parametric and non-parametric bootstrap methods.
}
\description{
A short description...
}
\details{
Additional details...
}
\examples{

rel_form <- Y ~ Yhat

inf_form <- Yhat ~ X1

dat <- simdat()

nboot <- 100

postpi_boot_ols(rel_form, inf_form, dat = dat, nboot)

}
