% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postpi_mi_ols.R
\name{postpi_mi_ols}
\alias{postpi_mi_ols}
\title{PostPI Linear Regression using Salerno et al. (2024) Multiple Imputation}
\usage{
postpi_mi_ols(
  rel_form,
  inf_form,
  dat,
  m = 100,
  rel_func = "lm",
  scale_sigmar = F
)
}
\arguments{
\item{rel_form}{A formula defining the relationship model. This should be
of the form Y ~ Yhat, where Y is the name of the column corresponding to
the observed outcome in the labeled data and Yhat is the name of the column
corresponding to the predicted outcome in the labeled data.}

\item{inf_form}{A formula defining the inference model. This should be of
the form Yhat ~ X, where Yhat is the name of the column corresponding to the
predicted outcome in the unlabeled data, and X generally corresponds to the
features of interest (e.g., X1 + X2).}

\item{dat}{data in the form of the simdat function}

\item{m}{number of imputations to be performed}

\item{rel_func}{relationship function model form (lm, gam, glm, etc)}

\item{scale_sigmar}{logical: scale sigma_r or not}
}
\value{
@returns A list of outputs: estimate of inference model parameters and corresponding standard error.
}
\description{
A short description...
}
\details{
Additional details...
}
\examples{

rel_form <- Y ~ Yhat

inf_form <- Yhat ~ X1

dat <- simdat()

m <- 100

postpi_mi_ols(rel_form, inf_form, dat = dat, m)

}
