% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_plusplus_logistic.R
\name{ppi_plusplus_logistic_est}
\alias{ppi_plusplus_logistic_est}
\title{PPI++ Logistic Regression Point Estimate}
\usage{
ppi_plusplus_logistic_est(
  X_l,
  Y_l,
  f_l,
  X_u,
  f_u,
  lhat = NULL,
  coord = NULL,
  opts = NULL,
  w_l = NULL,
  w_u = NULL
)
}
\arguments{
\item{X_l}{(matrix): n x p matrix of covariates in the labeled data.}

\item{Y_l}{(vector): n-vector of labeled outcomes.}

\item{f_l}{(vector): n-vector of predictions in the labeled data.}

\item{X_u}{(matrix): N x p matrix of covariates in the unlabeled data.}

\item{f_u}{(vector): N-vector of predictions in the unlabeled data.}

\item{lhat}{(float, optional): Power-tuning parameter.
The default value \code{NULL} will estimate the optimal value from data.
Setting \code{lhat = 1} recovers PPI with no power tuning.
Setting \code{lhat = 0} recovers the classical point estimate.}

\item{coord}{(int, optional): Coordinate for which to optimize \code{lhat}.
If \code{None}, it optimizes the total variance over all coordinates.
Must be in {1, ..., p} where p is the shape of the estimand.}

\item{opts}{(list, optional): Options to pass to the optimizer.
See ?optim for details.}

\item{w_l}{(vector, optional): n-vector of sample weights for the
labeled data.}

\item{w_u}{(vector, optional): N-vector of sample weights for the
unlabeled data.}
}
\value{
(vector): p-vector of prediction-powered point estimates of the
logistic regression coefficients.
}
\description{
Computes the prediction-powered point estimate of the logistic regression
coefficients.
}
\examples{

dat <- simdat_logistic()

form <- Y - Yhat ~ Xc

X_l <- model.matrix(form, data = dat[dat$set == "labeled",])

Y_l <- dat[dat$set == "labeled", all.vars(form)[1]] |> matrix(ncol = 1)

f_l <- dat[dat$set == "labeled", all.vars(form)[2]] |> matrix(ncol = 1)

X_u <- model.matrix(form, data = dat[dat$set == "unlabeled",])

f_u <- dat[dat$set == "unlabeled", all.vars(form)[2]] |> matrix(ncol = 1)

ppi_plusplus_logistic_est(X_l, Y_l, f_l, X_u, f_u)

}
