% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_plusplus_ols.R
\name{ppi_plusplus_ols}
\alias{ppi_plusplus_ols}
\title{PPI++ OLS Estimator and Inference}
\usage{
ppi_plusplus_ols(
  X_l,
  Y_l,
  f_l,
  X_u,
  f_u,
  n,
  p,
  N,
  alpha = 0.05,
  alternative = "two-sided",
  lhat = NULL,
  coord = NULL,
  w_l = NULL,
  w_u = NULL
)
}
\arguments{
\item{X_l}{(matrix): n x p matrix of covariates in the labeled data.}

\item{Y_l}{(vector): n-vector of labeled outcomes.}

\item{f_l}{(vector): n-vector of predictions in the labeled data.}

\item{X_u}{(matrix): N x p matrix of covariates in the unlabeled data.}

\item{f_u}{(vector): N-vector of predictions in the unlabeled data.}

\item{n}{(int): Number of labeled observations.}

\item{p}{(int): Number of covariates (features) of interest.}

\item{N}{(int): Number of unlabeled observations.}

\item{alpha}{(float): Significance level in [0,1]}

\item{alternative}{(string): Alternative hypothesis, either 'two-sided',
'larger' or 'smaller'.}

\item{lhat}{(float, optional): Power-tuning parameter.
The default value \code{NULL} will estimate the optimal value from data.
Setting \code{lhat = 1} recovers PPI with no power tuning.
Setting \code{lhat = 0} recovers the classical point estimate.}

\item{coord}{(int, optional): Coordinate for which to optimize \code{lhat}.
If \code{None}, it optimizes the total variance over all coordinates.
Must be in {1, ..., p} where p is the shape of the estimand.}

\item{w_l}{(vector, optional): n-vector of sample weights for the
labeled data.}

\item{w_u}{(vector, optional): N-vector of sample weights for the
unlabeled data.}
}
\value{
(list): A list containing the following:

\describe{
\item{est}{(vector): p-vector of PPI++ OLS coefficient estimates.}
\item{se}{(vector): p-vector of standard errors of the coefficients.}
}
}
\description{
Computes the prediction-powered estimator and confidence interval for the
OLS coefficients using the PPI++ algorithm.
}
\examples{

dat <- simdat()

form <- Y - Yhat ~ X1

X_l <- model.matrix(form, data = dat[dat$set == "tst",])

Y_l <- dat[dat$set == "tst", all.vars(form)[1]] |> matrix(ncol = 1)

f_l <- dat[dat$set == "tst", all.vars(form)[2]] |> matrix(ncol = 1)

X_u <- model.matrix(form, data = dat[dat$set == "val",])

f_u <- dat[dat$set == "val", all.vars(form)[2]] |> matrix(ncol = 1)

n <- nrow(X_l)

p <- ncol(X_l)

N <- nrow(X_u)

ppi_plusplus_ols(X_l, Y_l, f_l, X_u, f_u, n, p, N)

}
