% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_plusplus_quantile.R
\name{ppi_plusplus_quantile}
\alias{ppi_plusplus_quantile}
\title{PPI++ Quantile Estimation}
\usage{
ppi_plusplus_quantile(
  Y_l,
  f_l,
  f_u,
  q,
  alpha = 0.05,
  exact_grid = FALSE,
  w_l = NULL,
  w_u = NULL
)
}
\arguments{
\item{Y_l}{(vector): Gold-standard labels.}

\item{f_l}{(vector): Predictions corresponding to the gold-standard labels.}

\item{f_u}{(vector): Predictions corresponding to the unlabeled data.}

\item{q}{(float): Quantile to estimate. Must be in the range (0, 1).}

\item{alpha}{(float, optional): Error level; the confidence interval will
target a coverage of 1 - alpha. Must be in the range (0, 1).}

\item{exact_grid}{(bool, optional): Whether to use the exact grid of values
(TRUE) or a linearly spaced grid of 5000 values (FALSE).}

\item{w_l}{(vector, optional): Sample weights for the labeled data set.}

\item{w_u}{(vector, optional): Sample weights for the unlabeled data set.}
}
\value{
(list): Lower and upper bounds of the prediction-powered confidence  # Fix this to be estimate and sd
interval for the quantile.
}
\description{
Computes the prediction-powered confidence interval for the quantile.
}
\examples{

dat <- simdat()

form <- Y - f ~ X1

Y_l <- dat[dat$set == "labeled", all.vars(form)[1]] |> matrix(ncol = 1)

f_l <- dat[dat$set == "labeled", all.vars(form)[2]] |> matrix(ncol = 1)

f_u <- dat[dat$set == "unlabeled", all.vars(form)[2]] |> matrix(ncol = 1)

ppi_plusplus_quantile(Y_l, f_l, f_u, q = 0.5)

}
