% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_plusplus_quantile.R
\name{ppi_plusplus_quantile_est}
\alias{ppi_plusplus_quantile_est}
\title{PPI++ Quantile Estimation Point Estimate}
\usage{
ppi_plusplus_quantile_est(
  Y_l,
  f_l,
  f_u,
  q,
  exact_grid = FALSE,
  w_l = NULL,
  w_u = NULL
)
}
\arguments{
\item{Y_l}{(vector): Gold-standard labels.}

\item{f_l}{(vector): Predictions corresponding to the gold-standard labels.}

\item{f_u}{(vector): Predictions corresponding to the unlabeled data.}

\item{q}{(float): Quantile to estimate.}

\item{exact_grid}{(bool, optional): Whether to compute the exact solution
(TRUE) or an approximate solution based on a linearly spaced grid of 5000
values (FALSE).}

\item{w_l}{(vector, optional): Sample weights for the labeled data set.}

\item{w_u}{(vector, optional): Sample weights for the unlabeled data set.}
}
\value{
(float): Prediction-powered point estimate of the quantile.
}
\description{
Computes the prediction-powered point estimate of the quantile.
}
\examples{

dat <- simdat()

form <- Y - Yhat ~ X1

Y_l <- dat[dat$set == "tst", all.vars(form)[1]] |> matrix(ncol = 1)

f_l <- dat[dat$set == "tst", all.vars(form)[2]] |> matrix(ncol = 1)

f_u <- dat[dat$set == "val", all.vars(form)[2]] |> matrix(ncol = 1)

ppi_plusplus_quantile_est(Y_l, f_l, f_u, q = 0.5)

}
