% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.ipd}
\alias{print.ipd}
\title{Print ipd fit}
\usage{
\method{print}{ipd}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ipd}.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print ipd fit
}
