% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.ipd}
\alias{print.summary.ipd}
\title{Print summary.ipd}
\usage{
\method{print}{summary.ipd}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.ipd}.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print summary.ipd
}
