% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pspa.R
\name{sim_data_y}
\alias{sim_data_y}
\title{Simulate the data for testing the functions}
\usage{
sim_data_y(r = 0.9, binary = FALSE)
}
\arguments{
\item{r}{imputation correlation}

\item{binary}{simulate binary outcome or not}
}
\value{
simulated data
}
\description{
\code{sim_data_y} for simulation with ML-predicted Y
}
