% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.ipd}
\alias{summary.ipd}
\title{Summarize ipd fit}
\usage{
\method{summary}{ipd}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ipd}.}

\item{...}{Ignored.}
}
\value{
An object of class \code{summary.ipd} containing:
\describe{
\item{call}{The model formula.}
\item{coefficients}{A glm-style table of estimates, SE, z, p.}
\item{method}{Which IPD method was used.}
\item{model}{Which downstream model was fitted.}
\item{intercept}{Logical; whether an intercept was included.}
}
}
\description{
Summarize ipd fit
}
