% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tidy.ipd}
\alias{tidy.ipd}
\title{Tidy an ipd fit}
\usage{
\method{tidy}{ipd}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ipd}.}

\item{...}{Ignored.}
}
\value{
A \code{\link[tibble]{tibble}} with columns
\code{term, estimate, std.error, conf.low, conf.high}.
}
\description{
Tidy an ipd fit
}
\examples{

dat <- simdat()

fit <- ipd(Y - f ~ X1, method = "pspa", model = "ols",

    data = dat, label = "set_label")

tidy(fit)

}
