% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipd.R
\name{ipd}
\alias{ipd}
\title{Valid and Efficient Inference on Predicted Data (IPD) using State-of-the-Art
Methods}
\usage{
ipd(
  formula,
  method,
  model,
  data,
  label_index = NULL,
  df_unlabeled = NULL,
  seed = NULL,
  alpha = 0.05,
  alternative = "two-sided",
  ...
)
}
\arguments{
\item{formula}{(formula): an argument of the form Y - Yhat ~ X, where Y is the
name of the column corresponding to the observed outcome in the labeled data,
Yhat is the name of the column corresponding to the predicted outcome in both
labeled and unlabeled data, and X corresponds to the features of interest
(i.e., X = X1 + X2 + ... + Xp).}

\item{method}{(string): Method to be used. The available methods include
'postpi' (PostPI), 'ppi' (PPI), 'ppi_plusplus' (PPI++), and 'popinf' (POP-Inf).}

\item{model}{(string): Type of the regression model to be fit or
estimand to be calculated.
This can take on values: 'ols' (for continuous  variable), 'logistic'
(for binary variable), and estimands: 'mean' and 'quantile'.
Note that, not all method-model combinations are currently available. For newer
method-model additions, please use the development version from the github
repo: awanafiaz/IPD.}

\item{data}{(data.frame): either a stacked data frame with a specific column identifying the
labeled and unlabeled data sets, or only the labeled data set.
The stacked data frame (or the labeled data set) consists of observed outcomes
for labeled data set (Y), the predicted outcomes (Yhat) for both labeled and
unlabeled data sets, and the features (X).}

\item{label_index}{(string/int/logical, optional): The column name for the
indexing variable that identifies the labeled and unlabeled data sets. This
column must have the following pairs of values for labeled and unlabeled data:
string: 'labeled' and 'unlabeled', or integer: 1 (labeled) and 0 (unlabeled), or
logical: TRUE (labeled) and FALSE (unlabled)).}

\item{df_unlabeled}{(data.frame, optional): a data frame containing only the
unlabeled data set. Specify this argument ONLY if the data provided is not
already stacked. This data frame consists of the predicted outcomes (Yhat)
for the unlabeled data and the features (X). Specifying both 'label_index'
and 'df_unlabeled' arguments will result in an error message.}

\item{seed}{(int, optional): an integer value provided as seed.}

\item{alpha}{(float, optional): specified level of significance of the the
confidence interval; defaults to 0.05.}

\item{alternative}{(string, optional): Specify the alternative hypothesis,
must be one of "two.sided", "greater" or "less"; defaults to "two.sided".}

\item{...}{further arguments passed to or from specific methods. See details.}
}
\value{
a summary of model output.
}
\description{
A complete wrapper function with necessary defaults that calls
on each individual post-prediction inference based methods (postpi, ppi,
ppi++, popinf, etc) along with a specified model argument and
returns a summary of outputs.
}
\details{
Note that, going by the author-specified conventions, we denote
the methods as following:
(a) postpi       = Wang et al. (2020) post prediction inference (PostPI) method
(b) ppi          = Angelopoulos et al. (2023) prediction-powered inference
(PPI) method
(c) ppi_plusplus = Angelopoulos et al. (2023) PPI++ method
(d) popinf       = Miao et al. (2023) Assumption-lean and Data-adaptive
Post-Prediction Inference method

Since the PostPI methods (requiring a relationship model) are
the only ones to deviate from the 'rectifier' model framework (used in PPI,
PPI++, and POP-Inf), we will create the relationship model formula internally
when calling the PostPI methods under the assumption that the relationship
model arguments can be recovered from the rectifier model.
}
\examples{
# example code


}
