% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postpi_analytic_ols.R
\name{postpi_analytic_ols}
\alias{postpi_analytic_ols}
\title{PostPI Linear Regression using Wang et al. (2020) Analytic Correction}
\usage{
postpi_analytic_ols(X_l, Y_l, f_l, X_u, f_u, X_in_rel = FALSE)
}
\arguments{
\item{X_l}{(matrix): n x p matrix of covariates in the labeled data.}

\item{Y_l}{(vector): n-vector of labeled outcomes.}

\item{f_l}{(vector): n-vector of predictions in the labeled data.}

\item{X_u}{(matrix): N x p matrix of covariates in the unlabeled data.}

\item{f_u}{(vector): N-vector of predictions in the unlabeled data.}

\item{X_in_rel}{(boolean, optional): Logical argument for whether or not to
include the features of interest in the relationship model.
Defaults to FALSE.}
}
\value{
A list of outputs: estimate of the inference model parameters and
corresponding standard error estimate.
}
\description{
A short description.
}
\details{
Additional details...
}
\examples{

rel_form <- Y ~ Yhat

inf_form <- Yhat ~ X1

dat <- simdat()

postpi_analytic_ols(rel_form, inf_form, dat = dat)

}
