% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_plusplus_logistic.R
\name{ppi_plusplus_logistic}
\alias{ppi_plusplus_logistic}
\title{PPI++ Logistic Regression}
\usage{
ppi_plusplus_logistic(
  X_l,
  Y_l,
  f_l,
  X_u,
  f_u,
  lhat = NULL,
  coord = NULL,
  opts = NULL,
  w_l = NULL,
  w_u = NULL
)
}
\arguments{
\item{X_l}{(matrix): n x p matrix of covariates in the labeled data.}

\item{Y_l}{(vector): n-vector of labeled outcomes.}

\item{f_l}{(vector): n-vector of predictions in the labeled data.}

\item{X_u}{(matrix): N x p matrix of covariates in the unlabeled data.}

\item{f_u}{(vector): N-vector of predictions in the unlabeled data.}

\item{lhat}{(float, optional): Power-tuning parameter (see
\url{https://arxiv.org/abs/2311.01453}). The default value, \code{NULL},
will estimate the optimal value from the data. Setting \code{lhat = 1}
recovers PPI with no power tuning, and setting \code{lhat = 0} recovers
the classical point estimate.}

\item{coord}{(int, optional): Coordinate for which to optimize
\code{lhat = 1}. If \code{NULL}, it optimizes the total variance over all
coordinates. Must be in (1, ..., d) where d is the dimension of the estimand.}

\item{opts}{(list, optional): Options to pass to the optimizer.
See ?optim for details.}

\item{w_l}{(ndarray, optional): Sample weights for the labeled data set.
Defaults to a vector of ones.}

\item{w_u}{(ndarray, optional): Sample weights for the unlabeled
data set. Defaults to a vector of ones.}
}
\value{
(list): A list containing the following:

\describe{
\item{est}{(vector): vector of PPI++ logistic regression coefficient
estimates.}
\item{se}{(vector): vector of standard errors of the coefficients.}
\item{lambda}{(float): estimated power-tuning parameter.}
\item{rectifier_est}{(vector): vector of the rectifier logistic
regression coefficient estimates.}
\item{var_u}{(matrix): covariance matrix for the gradients in the
unlabeled data.}
\item{var_l}{(matrix): covariance matrix for the gradients in the
labeled data.}
\item{grads}{(matrix): matrix of gradients for the
labeled data.}
\item{grads_hat_unlabeled}{(matrix): matrix of predicted gradients for
the unlabeled data.}
\item{grads_hat}{(matrix): matrix of predicted gradients for the
labeled data.}
\item{inv_hessian}{(matrix): inverse Hessian matrix.}
}
}
\description{
Helper function for PPI++ logistic regression
}
\details{
PPI++: Efficient Prediction Powered Inference (Angelopoulos et al., 2023)
\url{https://arxiv.org/abs/2311.01453}
}
\examples{

dat <- simdat(model = "logistic")

form <- Y - f ~ X1

X_l <- model.matrix(form, data = dat[dat$set == "labeled",])

Y_l <- dat[dat$set == "labeled", all.vars(form)[1]] |> matrix(ncol = 1)

f_l <- dat[dat$set == "labeled", all.vars(form)[2]] |> matrix(ncol = 1)

X_u <- model.matrix(form, data = dat[dat$set == "unlabeled",])

f_u <- dat[dat$set == "unlabeled", all.vars(form)[2]] |> matrix(ncol = 1)

ppi_plusplus_logistic(X_l, Y_l, f_l, X_u, f_u)

}
