% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pspa.R
\name{pspa_y}
\alias{pspa_y}
\title{PSPA M-Estimation for ML-predicted labels}
\usage{
pspa_y(
  X_l = NA,
  X_u = NA,
  Y_l,
  f_l,
  f_u,
  alpha = 0.05,
  weights = NA,
  quant = NA,
  intercept = FALSE,
  method
)
}
\arguments{
\item{X_l}{Array or data.frame containing observed covariates in
labeled data.}

\item{X_u}{Array or data.frame containing observed or predicted
covariates in unlabeled data.}

\item{Y_l}{Array or data.frame of observed outcomes in
labeled data.}

\item{f_l}{Array or data.frame of predicted outcomes in
labeled data.}

\item{f_u}{Array or data.frame of predicted outcomes in
unlabeled data.}

\item{alpha}{Specifies the confidence level as 1 - alpha for
confidence intervals.}

\item{weights}{weights vector PSPA linear regression (d-dimensional, where
d equals the number of covariates).}

\item{quant}{quantile for quantile estimation}

\item{intercept}{Boolean indicating if the input covariates' data contains
the intercept (TRUE if the input data contains)}

\item{method}{indicates the method to be used for M-estimation.
Options include "mean", "quantile", "ols", "logistic", and "poisson".}
}
\value{
A summary table presenting point estimates, standard error,
confidence intervals (1 - alpha), P-values, and weights.
}
\description{
\code{pspa_y} function conducts post-prediction M-Estimation.
}
