% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pspa.R
\name{Sigma_cal}
\alias{Sigma_cal}
\title{Variance-covariance matrix of the estimation equation}
\usage{
Sigma_cal(
  X_l,
  X_u,
  Y_l,
  f_l,
  f_u,
  w,
  theta,
  quant = NA,
  method = c("ols", "quantile", "mean", "logistic", "poisson")
)
}
\arguments{
\item{X_l}{Array or data.frame containing observed covariates in
labeled data.}

\item{X_u}{Array or data.frame containing observed or predicted
covariates in unlabeled data.}

\item{Y_l}{Array or data.frame of observed outcomes in labeled data.}

\item{f_l}{Array or data.frame of predicted outcomes in labeled data.}

\item{f_u}{Array or data.frame of predicted outcomes in
unlabeled data.}

\item{w}{weights vector PSPA linear regression (d-dimensional, where
d equals the number of covariates).}

\item{theta}{parameter theta}

\item{quant}{quantile for quantile estimation}

\item{method}{indicates the method to be used for M-estimation.
Options include "mean", "quantile", "ols", "logistic", and "poisson".}
}
\value{
variance-covariance matrix of the estimation equation
}
\description{
\code{Sigma_cal} function for variance-covariance matrix of the
estimation equation
}
