% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdat.R
\name{simdat}
\alias{simdat}
\title{Data generation for OLS examples}
\usage{
simdat(n = c(300, 300, 300), beta1 = 1)
}
\arguments{
\item{n}{vector of size 3 indicating the sample size in the training,
labelled, and unlabelled data sets}

\item{beta1}{first regression coefficient (or, regression coefficient of
variable of interest for inference)}
}
\value{
A data frame containing 4 regressors, labelled outcome, predicted
outcome and a character variable indicating which data set the observation
belongs to (training, test/labeled, validation/unlabeled).
}
\description{
Data generation for OLS examples
}
\examples{

simdat(c(100, 100, 100), 1)

}
