% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdat.R
\name{simdat_logistic}
\alias{simdat_logistic}
\title{Data generation for logistic regression examples}
\usage{
simdat_logistic(n = c(300, 300, 300), betac = 1)
}
\arguments{
\item{n}{vector of size 3 indicating the sample size in the training,
labelled, and unlabelled data sets}

\item{betac}{first regression coefficient (or, regression coefficient of
variable of interest for inference)}
}
\value{
A data frame containing 4 regressors, labelled outcome, predicted
outcome and a character variable indicating which dat set the observartion
belongs to (training, test, validation).
}
\description{
Data generation for logistic regression examples
}
\examples{

simdat(c(100, 100, 100), 1)

}
